requirements.txt
transformers
torch
accelerate
openai-whisper
google-cloud-dialogflow
gtts
streamlit
